+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     - `keyType`: keys of the map
     - `valueType`: values of the map
 - Added support for `include_nulls` argument in `DataFrame.unpivot`.
+- Added support for `Dataframe.toJSON`
 
 #### Improvements